renamed component names as per guidelines
