+++ src/client/containers/EmptyQuizResultPage/EmptyQuizResultPage.styles.css
